Gael o2 emilie

+++ GaelO2/GaelO2/routes/api.php
@@ -12,13 +12,27 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+/*
+|--------------------------------------------------------------------------
+| Admin Routes
+|--------------------------------------------------------------------------
+|
+*/
+
+//User related Routes
+Route::middleware(['auth:api', 'refresh_token'])->post('users', 'UserController@createUser');
+Route::middleware(['auth:api', 'refresh_token'])->put('users/self/{id}', 'UserController@modifyUserIdentification');
+Route::middleware(['auth:api', 'refresh_token'])->delete('users/{id}', 'UserController@deleteUser');
+
+//Export DB Route (download binary doesn't support refresh token)
+Route::middleware('auth:api')->post('export-db', 'ExportDBController@exportDB');
+
+
 //Routes that need authentication and to be admin
 Route::middleware(['auth:api', 'admin', 'refresh_token'])->group(function () {
 
     //Users Routes
-    Route::post('users', 'UserController@createUser');
     Route::put('users/{id}', 'UserController@modifyUser');
-    Route::delete('users/{id}', 'UserController@deleteUser');
     Route::patch('users/{id}/reactivate', 'UserController@reactivateUser');
     Route::get('users/{id}/roles/{study?}', 'UserController@getRoles');
     Route::post('users/{id}/roles/{study}', 'UserController@createRole');
@@ -60,8 +74,13 @@ Route::middleware(['auth:api', 'admin', 'refresh_token'])->group(function () {
 
 });
 
-//Export DB Route
-Route::middleware('auth:api')->post('export-db', 'ExportDBController@exportDB');
+
+/*
+|--------------------------------------------------------------------------
+| Users Routes
+|--------------------------------------------------------------------------
+|
+*/
 
 //Routes that need authentication
 Route::middleware(['auth:api', 'refresh_token'])->group(function () {
@@ -76,16 +95,21 @@ Route::middleware(['auth:api', 'refresh_token'])->group(function () {
     Route::get('studies/{studyName}/patients', 'PatientController@getPatientFromStudy');
     Route::post('studies/{studyName}/import-patients', 'StudyController@importPatients');
 
-    //Patient Tree
-    //Route::get('studies/{studyName}/tree', 'StudyController@getStudyInfoForTree');
-
     //Visit Routes
     Route::post('studies/{studyName}/visit-groups/{visitGroupId}/visit-types/{visitTypeId}/visits', 'VisitController@createVisit');
-    Route::get('studies/{studyName}/visit-groups/{visitGroupId}/visit-types/{visitTypeId}/visits/{id}', 'VisitController@getVisit');
-    Route::get('studies/{studyName}/visit-groups/{visitGroupId}/visit-types/{visitTypeId}/visits/{id}/patients/{patientCode}', 'VisitController@getPatientVisit');
+    Route::get('visits/{id}', 'VisitController@getVisit');
+    Route::get('visits/{id}/patients/{patientCode}', 'VisitController@getPatientVisit');
 });
 
-//Mail Route
+
+/*
+|--------------------------------------------------------------------------
+| Public Routes
+|--------------------------------------------------------------------------
+|
+*/
+
+//Request Route
 Route::post('request', 'RequestController@sendRequest');
 
 //Login Route
